Show hopefully friendlier messages
Since relative imports don't work with the `import <>` syntax, suggest
importing the names directly. In these cases, the `i.module` is `None`
and the `i.level` has the number of levels down in the relative import,
e.g. `1` means the current directory.

If there's only one name, suggest:

    use 'import module.name' instead of 'from module import name'

This implements the enhancement proposed in #1.